requirements.txt
anndata==0.9.1
h5py==3.9.0
hnswlib==0.7.0
numba==0.57.1
numpy==1.24.3
pandas==1.3.5 
python==3.8.16
pytorch==1.13.1
pytorch-cuda==11.7
scanpy==1.9.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
